requirements.txt
stardist
pywin32
imutils
PyQt5-tools
opencv-python
pyinstaller